+++ packages/web-data/.resources.mjs
@@ -1,17 +1,31 @@
 export default [
   {
     path: '/onnxruntime-web/',
-    source: '../../node_modules/onnxruntime-web/dist/*.wasm',
+    source:
+      '../../node_modules/onnxruntime-web/dist/ort-wasm-simd-threaded.jsep.wasm',
     mime: 'application/wasm'
   },
   {
-    path: '/models/',
-    source: '../../bundle/models/small',
-    mime: 'application/octet-steam'
+    path: '/onnxruntime-web/',
+    source:
+      '../../node_modules/onnxruntime-web/dist/ort-wasm-simd-threaded.wasm',
+    mime: 'application/wasm'
+  },
+  {
+    path: '/onnxruntime-web/',
+    source:
+      '../../node_modules/onnxruntime-web/dist/ort-wasm-simd-threaded.jsep.mjs',
+    mime: 'text/javascript'
+  },
+  {
+    path: '/onnxruntime-web/',
+    source:
+      '../../node_modules/onnxruntime-web/dist/ort-wasm-simd-threaded.mjs',
+    mime: 'text/javascript'
   },
   {
     path: '/models/',
-    source: '../../bundle/models/medium',
+    source: '../../bundle/models/*',
     mime: 'application/octet-steam'
   }
 ];