+++ packages/web/src/onnx.ts
@@ -1,57 +1,52 @@
 export { createOnnxSession, runOnnxSession };
 
 import ndarray, { NdArray } from 'ndarray';
-import * as ort from 'onnxruntime-web';
-import { loadAsUrl } from './resource';
-import { simd, threads } from './feature-detect';
+import type ORT from 'onnxruntime-web';
+
+import * as ort_cpu from 'onnxruntime-web';
+import * as ort_gpu from 'onnxruntime-web/webgpu';
+
+import * as caps from './capabilities';
 import { Config } from './schema';
+import { loadAsUrl, resolveChunkUrls } from './resource';
 
 async function createOnnxSession(model: any, config: Config) {
-  const capabilities = {
-    simd: await simd(),
-    threads: await threads(),
-    numThreads: navigator.hardwareConcurrency ?? 4,
-    // @ts-ignore
-    webgpu: navigator.gpu !== undefined
-  };
+  const useWebGPU = config.device === 'gpu' && (await caps.webgpu());
+  // BUG: proxyToWorker is not working for WASM/CPU Backend for now
+  const proxyToWorker = useWebGPU && config.proxyToWorker;
+  const executionProviders = [useWebGPU ? 'webgpu' : 'wasm'];
+  const ort = useWebGPU ? ort_gpu : ort_cpu;
+
   if (config.debug) {
-    console.debug('Capabilities:', capabilities);
+    console.debug('\tUsing WebGPU:', useWebGPU);
+    console.debug('\tProxy to Worker:', proxyToWorker);
+
     ort.env.debug = true;
     ort.env.logLevel = 'verbose';
   }
 
-  ort.env.wasm.numThreads = capabilities.numThreads;
-  ort.env.wasm.simd = capabilities.simd;
-  ort.env.wasm.proxy = config.proxyToWorker;
+  ort.env.wasm.numThreads = caps.maxNumThreads();
+  ort.env.wasm.proxy = proxyToWorker;
+
+  // The path inside the resource bundle
+  const baseFilePath = useWebGPU
+    ? '/onnxruntime-web/ort-wasm-simd-threaded.jsep'
+    : '/onnxruntime-web/ort-wasm-simd-threaded';
+
+  const wasmPath = await loadAsUrl(`${baseFilePath}.wasm`, config);
+  const mjsPath = await loadAsUrl(`${baseFilePath}.mjs`, config);
 
   ort.env.wasm.wasmPaths = {
-    'ort-wasm-simd-threaded.wasm':
-      capabilities.simd && capabilities.threads
-        ? await loadAsUrl(
-            '/onnxruntime-web/ort-wasm-simd-threaded.wasm',
-            config
-          )
-        : undefined,
-    'ort-wasm-simd.wasm':
-      capabilities.simd && !capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm-simd.wasm', config)
-        : undefined,
-    'ort-wasm-threaded.wasm':
-      !capabilities.simd && capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm-threaded.wasm', config)
-        : undefined,
-    'ort-wasm.wasm':
-      !capabilities.simd && !capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm.wasm', config)
-        : undefined
+    mjs: mjsPath,
+    wasm: wasmPath
   };
 
   if (config.debug) {
     console.debug('ort.env.wasm:', ort.env.wasm);
   }
 
-  const ort_config: ort.InferenceSession.SessionOptions = {
-    executionProviders: ['wasm'],
+  const ort_config: ORT.InferenceSession.SessionOptions = {
+    executionProviders: executionProviders,
     graphOptimizationLevel: 'all',
     executionMode: 'parallel',
     enableCpuMemArena: true
@@ -60,7 +55,7 @@ async function createOnnxSession(model: any, config: Config) {
   const session = await ort.InferenceSession.create(model, ort_config).catch(
     (e: any) => {
       throw new Error(
-        `Failed to create session: ${e}. Please check if the publicPath is set correctly.`
+        `Failed to create session: "${e}". Please check if the publicPath is set correctly.`
       );
     }
   );
@@ -70,8 +65,12 @@ async function createOnnxSession(model: any, config: Config) {
 async function runOnnxSession(
   session: any,
   inputs: [string, NdArray<Float32Array>][],
-  outputs: [string]
+  outputs: [string],
+  config: Config
 ) {
+  const useWebGPU = config.device === 'gpu' && (await caps.webgpu());
+  const ort = useWebGPU ? ort_gpu : ort_cpu;
+
   const feeds: Record<string, any> = {};
   for (const [key, tensor] of inputs) {
     feeds[key] = new ort.Tensor(
@@ -82,9 +81,8 @@ async function runOnnxSession(
   }
   const outputData = await session.run(feeds, {});
   const outputKVPairs: NdArray<Float32Array>[] = [];
-
   for (const key of outputs) {
-    const output: ort.Tensor = outputData[key];
+    const output: ORT.Tensor = outputData[key];
     const shape: number[] = output.dims as number[];
     const data: Float32Array = output.data as Float32Array;
     const tensor = ndarray(data, shape);