+++ packages/web/CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.6]
+
+### Changed
+
+- Upgrade onnx-runtime
+
+## [1.5.0]
+
+### Added
+
+- Added option to execute on gpu (webgpu) and cpu. Enable by setting `device: "gpu"`
+
+- Added isnet model for webgpu
+
+## [1.4.5]
+
+### Added
+
+- Added ThirdPartyLicenses.json
+
 ## [1.4.0]
 
 ### Added